requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioredis==1.3.1
anyio==3.6.1; python_full_version >= '3.6.2'
async-timeout==4.0.2; python_version >= '3.6'
bcrypt==4.0.0
certifi==2022.9.14; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
cryptography==38.0.1
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
elasticsearch==7.13.0
fastapi-cache2[redis]==0.1.9
fastapi==0.85.0
gql[aiohttp]==0.5.0
graphql-core==2.0.0
gunicorn==20.1.0
h11==0.13.0; python_version >= '3.6'
hiredis==2.0.0; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
packaging==21.3; python_version >= '3.6'
passlib[bcrypt]==1.7.4
pendulum==2.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
promise==2.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2; python_version >= '3.7'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-jose[cryptography]==3.3.0
python-json-logger==2.0.4
pytzdata==2020.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
redis==4.3.4
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
rx==3.2.0; python_version >= '3.6'
setuptools==65.3.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.20.4; python_version >= '3.7'
typing-extensions==4.3.0; python_version < '3.10'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
uvicorn==0.18.3
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
